Add Repository Discussions using GraphQL API
