+++ github/TeamDiscussion.py
@@ -23,6 +23,7 @@
 # Copyright 2023 Trim21 <trim21.me@gmail.com>                                  #
 # Copyright 2024 Enrico Minack <github@enrico.minack.dev>                      #
 # Copyright 2024 Jirka Borovec <6035284+Borda@users.noreply.github.com>        #
+# Copyright 2024 Enrico Minack <github@enrico.minack.dev>                      #
 #                                                                              #
 # This file is part of PyGithub.                                               #
 # http://pygithub.readthedocs.io/                                              #
@@ -44,15 +45,13 @@
 
 from __future__ import annotations
 
-from datetime import datetime
 from typing import Any
 
-import github.GithubObject
-import github.NamedUser
-from github.GithubObject import Attribute, CompletableGithubObject, NotSet
+from github.DiscussionBase import DiscussionBase
+from github.GithubObject import Attribute, NotSet
 
 
-class TeamDiscussion(CompletableGithubObject):
+class TeamDiscussion(DiscussionBase):
     """
     This class represents TeamDiscussions.
 
@@ -62,41 +61,15 @@ class TeamDiscussion(CompletableGithubObject):
     """
 
     def _initAttributes(self) -> None:
-        self._author: Attribute[github.NamedUser.NamedUser] = NotSet
-        self._body: Attribute[str] = NotSet
-        self._body_html: Attribute[str] = NotSet
+        super()._initAttributes()
         self._body_version: Attribute[str] = NotSet
         self._comments_count: Attribute[int] = NotSet
         self._comments_url: Attribute[str] = NotSet
-        self._created_at: Attribute[datetime] = NotSet
         self._html_url: Attribute[str] = NotSet
-        self._last_edited_at: Attribute[datetime] = NotSet
         self._node_id: Attribute[str] = NotSet
-        self._number: Attribute[int] = NotSet
         self._pinned: Attribute[bool] = NotSet
         self._private: Attribute[bool] = NotSet
         self._team_url: Attribute[str] = NotSet
-        self._title: Attribute[str] = NotSet
-        self._updated_at: Attribute[datetime] = NotSet
-        self._url: Attribute[str] = NotSet
-
-    def __repr__(self) -> str:
-        return self.get__repr__({"number": self._number.value, "title": self._title.value})
-
-    @property
-    def author(self) -> github.NamedUser.NamedUser:
-        self._completeIfNotSet(self._author)
-        return self._author.value
-
-    @property
-    def body(self) -> str:
-        self._completeIfNotSet(self._body)
-        return self._body.value
-
-    @property
-    def body_html(self) -> str:
-        self._completeIfNotSet(self._body_html)
-        return self._body_html.value
 
     @property
     def body_version(self) -> str:
@@ -114,29 +87,14 @@ class TeamDiscussion(CompletableGithubObject):
         return self._comments_url.value
 
     @property
-    def created_at(self) -> datetime:
-        self._completeIfNotSet(self._created_at)
-        return self._created_at.value
-
-    @property
     def html_url(self) -> str:
         self._completeIfNotSet(self._html_url)
         return self._html_url.value
 
     @property
-    def last_edited_at(self) -> datetime:
-        self._completeIfNotSet(self._last_edited_at)
-        return self._last_edited_at.value
-
-    @property
     def node_id(self) -> str:
         self._completeIfNotSet(self._node_id)
         return self._node_id.value
-
-    @property
-    def number(self) -> int:
-        self._completeIfNotSet(self._number)
-        return self._number.value
 
     @property
     def pinned(self) -> bool:
@@ -153,53 +111,21 @@ class TeamDiscussion(CompletableGithubObject):
         self._completeIfNotSet(self._team_url)
         return self._team_url.value
 
-    @property
-    def title(self) -> str:
-        self._completeIfNotSet(self._title)
-        return self._title.value
-
-    @property
-    def updated_at(self) -> datetime:
-        self._completeIfNotSet(self._updated_at)
-        return self._updated_at.value
-
-    @property
-    def url(self) -> str:
-        self._completeIfNotSet(self._url)
-        return self._url.value
-
     def _useAttributes(self, attributes: dict[str, Any]) -> None:
-        if "author" in attributes:  # pragma no branch
-            self._author = self._makeClassAttribute(github.NamedUser.NamedUser, attributes["author"])
-        if "body" in attributes:  # pragma no branch
-            self._body = self._makeStringAttribute(attributes["body"])
-        if "body_html" in attributes:  # pragma no branch
-            self._body_html = self._makeStringAttribute(attributes["body_html"])
+        super()._useAttributes(attributes)
         if "body_version" in attributes:  # pragma no branch
             self._body_version = self._makeStringAttribute(attributes["body_version"])
         if "comments_count" in attributes:  # pragma no branch
             self._comments_count = self._makeIntAttribute(attributes["comments_count"])
         if "comments_url" in attributes:  # pragma no branch
             self._comments_url = self._makeStringAttribute(attributes["comments_url"])
-        if "created_at" in attributes:  # pragma no branch
-            self._created_at = self._makeDatetimeAttribute(attributes["created_at"])
         if "html_url" in attributes:  # pragma no branch
             self._html_url = self._makeStringAttribute(attributes["html_url"])
-        if "last_edited_at" in attributes:  # pragma no branch
-            self._last_edited_at = self._makeDatetimeAttribute(attributes["last_edited_at"])
         if "node_id" in attributes:  # pragma no branch
             self._node_id = self._makeStringAttribute(attributes["node_id"])
-        if "number" in attributes:  # pragma no branch
-            self._number = self._makeIntAttribute(attributes["number"])
         if "pinned" in attributes:  # pragma no branch
             self._pinned = self._makeBoolAttribute(attributes["pinned"])
         if "private" in attributes:  # pragma no branch
             self._private = self._makeBoolAttribute(attributes["private"])
         if "team_url" in attributes:  # pragma no branch
             self._team_url = self._makeStringAttribute(attributes["team_url"])
-        if "title" in attributes:
-            self._title = self._makeStringAttribute(attributes["title"])
-        if "updated_at" in attributes:  # pragma no branch
-            self._updated_at = self._makeDatetimeAttribute(attributes["updated_at"])
-        if "url" in attributes:  # pragma no branch
-            self._url = self._makeStringAttribute(attributes["url"])
